+++ docs/source/changes.md
@@ -26,6 +26,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`579` fixes an interaction with `--pdb` and `--trace` and task that return. The
   debugging modes swallowed the return and `None` was returned. Closes {issue}`574`.
 - {pull}`581` simplifies the code for tracebacks and unpublishes some utility functions.
+- {pull}`585` implements two more import modes, `prepend` and `append`.
 
 ## 0.4.6 - 2024-03-13